fix: simplify trpc file structure
