--- src/server/trpc/init.ts
@@ -1,25 +1,0 @@
-import { createTRPCContext } from '@/server/trpc';
-import { initTRPC } from '@trpc/server';
-import { headers } from 'next/headers';
-import { cache } from 'react';
-import superjson from 'superjson';
-
-export const createContext = cache(async () => {
-  const heads = new Headers(await headers());
-  heads.set('x-trpc-source', 'rsc');
-
-  return createTRPCContext({
-    headers: heads,
-  });
-});
-
-const t = initTRPC.create({
-  /**
-   * @see https://trpc.io/docs/server/data-transformers
-   */
-  transformer: superjson,
-});
-// Base router and procedure helpers
-export const createTRPCRouter = t.router;
-export const createCallerFactory = t.createCallerFactory;
-export const baseProcedure = t.procedure;
